[fs] nfsd: fix pNFS return on close semantics
Message-id: <1466453769-4933-54-git-send-email-bfields@redhat.com>
Patchwork-id: 153738
O-Subject: [Patch RHEL7 fs 053/143] nfsd: fix pNFS return on close semantics
Bugzilla: 1344797
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Sachin Bhamare <sachin.bhamare@primarydata.com>

For the sake of forgetful clients, the server should return the layouts
to the file system on 'last close' of a file (assuming that there are no
delegations outstanding to that particular client) or on delegreturn
(assuming that there are no opens on a file from that particular
client).

In theory the information is all there in current data structures, but
it's not efficiently available; nfs4_file->fi_ref includes references on
the file across all clients, but we need a per-(client, file) count.
Walking through lots of stateid's to calculate this on each close or
delegreturn would be painful.

This patch introduces infrastructure to maintain per-client opens and
delegation counters on a per-file basis.

[hch: ported to the mainline pNFS support, merged various fixes from Jeff]
Signed-off-by: Sachin Bhamare <sachin.bhamare@primarydata.com>
Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>

(cherry picked from commit 8287f009bd95a5e548059dba62a67727bb9549cd)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>