[fs] nfsd: serialize layout stateid morphing operations
Message-id: <1466453769-4933-120-git-send-email-bfields@redhat.com>
Patchwork-id: 153779
O-Subject: [Patch RHEL7 fs 119/143] nfsd: serialize layout stateid morphing operations
Bugzilla: 1344797
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

In order to allow the client to make a sane determination of what
happened with racing LAYOUTGET/LAYOUTRETURN/CB_LAYOUTRECALL calls, we
must ensure that the seqids return accurately represent the order of
operations. The simplest way to do that is to ensure that operations on
a single stateid are serialized.

This patch adds a mutex to the layout stateid, and locks it when
checking the layout stateid's seqid. The mutex is held over the entire
operation and released after the seqid is bumped.

Note that in the case of CB_LAYOUTRECALL we must move the increment of
the seqid and setting into a new cb "prepare" operation. The lease
infrastructure will call the lm_break callback with a spinlock held, so
and we can't take the mutex in that codepath.

Cc: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
(cherry picked from commit cc8a55320b5f1196bee5bd14e4bb2ebd3b983317)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>