[fs] nfsd: ensure that seqid morphing operations are atomic wrt to copies
Message-id: <1466453769-4933-121-git-send-email-bfields@redhat.com>
Patchwork-id: 153799
O-Subject: [Patch RHEL7 fs 120/143] nfsd: ensure that seqid morphing operations are atomic wrt to copies
Bugzilla: 1344797
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

Bruce points out that the increment of the seqid in stateids is not
serialized in any way, so it's possible for racing calls to bump it
twice and end up sending the same stateid. While we don't have any
reports of this problem it _is_ theoretically possible, and could lead
to spurious state recovery by the client.

In the current code, update_stateid is always followed by a memcpy of
that stateid, so we can combine the two operations. For better
atomicity, we add a spinlock to the nfs4_stid and hold that when bumping
the seqid and copying the stateid.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
(cherry picked from commit 9767feb2c64b29775f1ea683130b44f95f67d169)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>