[fs] nfsd: fix race with open / open upgrade stateids
Message-id: <1466453769-4933-124-git-send-email-bfields@redhat.com>
Patchwork-id: 153787
O-Subject: [Patch RHEL7 fs 123/143] nfsd: fix race with open / open upgrade stateids
Bugzilla: 1344797
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Andrew Elble <aweits@rit.edu>

We observed multiple open stateids on the server for files that
seemingly should have been closed.

nfsd4_process_open2() tests for the existence of a preexisting
stateid. If one is not found, the locks are dropped and a new
one is created. The problem is that init_open_stateid(), which
is also responsible for hashing the newly initialized stateid,
doesn't check to see if another open has raced in and created
a matching stateid. This fix is to enable init_open_stateid() to
return the matching stateid and have nfsd4_process_open2()
swap to that stateid and switch to the open upgrade path.
In testing this patch, coverage to the newly created
path indicates that the race was indeed happening.

Signed-off-by: Andrew Elble <aweits@rit.edu>
Reviewed-by: Jeff Layton <jlayton@poochiereds.net>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
(cherry picked from commit 7fc0564e3a8d16df096f48c9c6425ba84d945c6e)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>