[fs] lib/vsprintf.c: fix potential NULL deref in hex_string
Message-id: <7d8ff322b1dbefa67580cf8f76487c58f324081d.1467175229.git.bcodding@redhat.com>
Patchwork-id: 154255
O-Subject: [RHEL7 fs PATCH 2/7] lib/vsprintf.c: fix potential NULL deref in hex_string
Bugzilla: 1305092
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Rasmus Villemoes <linux@rasmusvillemoes.dk>

The helper hex_string() is broken in two ways.  First, it doesn't
increment buf regardless of whether there is room to print, so callers
such as kasprintf() that try to probe the correct storage to allocate will
get a too small return value.  But even worse, kasprintf() (and likely
anyone else trying to find the size of the result) pass NULL for buf and 0
for size, so we also have end == NULL.  But this means that the end-1 in
hex_string() is (char*)-1, so buf < end-1 is true and we get a NULL
pointer deref.  I double-checked this with a trivial kernel module that
just did a kasprintf(GFP_KERNEL, "%14ph", "CrashBoomBang").

Nobody seems to be using %ph with kasprintf, but we might as well fix it
before it hits someone.

Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Acked-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1305092
Upstream-Id: 9c98f2359600386efd1c6adfabc5a04118a79798
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>