[fs] gfs2: Wait for iopen glock dequeues
Message-id: <1467230737-19262-3-git-send-email-rpeterso@redhat.com>
Patchwork-id: 154286
O-Subject: [RHEL7.3 PATCH 02/14] GFS2: Wait for iopen glock dequeues
Bugzilla: 1336011
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch changes every glock_dq for iopen glocks into a dq_wait.
This makes sure that iopen glocks do not outlive the inode itself.
In turn, that ensures that anyone trying to unlink the glock will
be able to find the inode when it receives a remote iopen callback.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

This is a RHEL7 crosswrite of the following upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=86d067a797d4e8546a7c92b985f31e8cd3ec39ad

rhbz#1336011
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>