[fs] gfs2: Release iopen glock in gfs2_create_inode error cases
Message-id: <1467230737-19262-4-git-send-email-rpeterso@redhat.com>
Patchwork-id: 154285
O-Subject: [RHEL7.3 PATCH 03/14] GFS2: Release iopen glock in gfs2_create_inode error cases
Bugzilla: 1336011
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Some error cases in gfs2_create_inode were not unlocking the iopen
glock, getting the reference count off. This adds the proper unlock.
The error logic in function gfs2_create_inode was also convoluted,
so this patch simplifies it. It also takes care of a bug in
which gfs2_qa_delete() was not called in an error case.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
This is a RHEL7 crosswrite of the following upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=783013c0f5c7263a31703b15aeebbac279b4d4fe

rhbz#1336011
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>