[fs] gfs2: Always use iopen glock for gl_deletes
Message-id: <1467230737-19262-5-git-send-email-rpeterso@redhat.com>
Patchwork-id: 154298
O-Subject: [RHEL7.3 PATCH 04/14] GFS2: Always use iopen glock for gl_deletes
Bugzilla: 1336011
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Before this patch, when function try_rgrp_unlink queued a glock for
delete_work to reclaim the space, it used the inode glock to do so.
That's different from the iopen callback which uses the iopen glock
for the same purpose. We should be consistent and always use the
iopen glock. This may also save us reference counting problems with
the inode glock, since clear_glock does an extra glock_put() for the
inode glock.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

This is a RHEL7 crosswrite of the following upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=5ea31bc0a6524b4fee8dc9ae8005d4a114a79812

rhbz#1336011
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>