[fs] gfs2: Check if iopen is held when deleting inode
Message-id: <1467230737-19262-8-git-send-email-rpeterso@redhat.com>
Patchwork-id: 154294
O-Subject: [RHEL7.3 PATCH 07/14] GFS2: Check if iopen is held when deleting inode
Bugzilla: 1336011
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch fixes an error condition in which an inode is partially
created in gfs2_create_inode() but then some error is discovered,
which causes it to fail and call iput() before the iopen glock is
created or held. In that case, gfs2_delete_inode would try to
unlock an iopen glock that doesn't yet exist. Therefore, we test
its holder (which must exist) for the HIF_HOLDER bit before trying
to dq it.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This is a RHEL7 crosswrite of the following upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=7508abc4bdac43dc87d2fdd31527063f72da7020

rhbz#1336011
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>