[fs] gfs2: Fix gfs2_lookup_by_inum lock inversion
Message-id: <1467230737-19262-12-git-send-email-rpeterso@redhat.com>
Patchwork-id: 154295
O-Subject: [RHEL7.3 PATCH 11/14] gfs2: Fix gfs2_lookup_by_inum lock inversion
Bugzilla: 1336011
RH-Acked-by: Andreas Gruenbacher <agruenba@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Andreas Gruenbacher <agruenba@redhat.com>

The current gfs2_lookup_by_inum takes the glock of a presumed inode
identified by block number, verifies that the block is indeed an inode,
and then instantiates and reads the new inode via gfs2_inode_lookup.

However, instantiating a new inode may block on freeing a previous
instance of that inode (__wait_on_freeing_inode), and freeing an inode
requires to take the glock already held, leading to lock inversion and
deadlock.

Fix this by first instantiating the new inode, then verifying that the
block is an inode (if required), and then reading in the new inode, all
in gfs2_inode_lookup.

If the block we are looking for is not an inode, we discard the new
inode via iget_failed, which marks inodes as bad and unhashes them.
Other tasks waiting on that inode will get back a bad inode back from
ilookup or iget_locked; in that case, retry the lookup.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Bob Peterson <rpeterso@redhat.com>

This is a RHEL7 crosswrite of the following upstream patch:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=3ce37b2cb4917674fa5b776e857dcea94c0e0835

rhbz#1336011
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>