[fs] nfsd: add SCSI layout support
Message-id: <a83df9448ddfaca7f6f623683d831dbe4597911b.1467719857.git.bcodding@redhat.com>
Patchwork-id: 154661
O-Subject: [RHEL7 fs PATCH v2 5/9] nfsd: add SCSI layout support
Bugzilla: 1305094
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Christoph Hellwig <hch@lst.de>

This is a simple extension to the block layout driver to use SCSI
persistent reservations for access control and fencing, as well as
SCSI VPD pages for device identification.

For this we need to pass the nfs4_client to the proc_getdeviceinfo method
to generate the reservation key, and add a new fence_client method
to allow for fence actions in the layout driver.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream-Id: f99d4fbdae6765d0bb4ed5441f6fa1f036122d59
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
---
 create mode 100644 Documentation/filesystems/nfs/pnfs-scsi-server.txt

Signed-off-by: Rafael Aquini <aquini@redhat.com>