[virt] kvm: x86: Add EOI exit bitmap inference
Message-id: <1467818058-27383-62-git-send-email-pbonzini@redhat.com>
Patchwork-id: 154750
O-Subject: [RHEL7.3 PATCH v3 61/77] KVM: x86: Add EOI exit bitmap inference
Bugzilla: 1347370
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In order to support a userspace IOAPIC interacting with an in kernel
APIC, the EOI exit bitmaps need to be configurable.

If the IOAPIC is in userspace (i.e. the irqchip has been split), the
EOI exit bitmaps will be set whenever the GSI Routes are configured.
In particular, for the low MSI routes are reservable for userspace
IOAPICs. For these MSI routes, the EOI Exit bit corresponding to the
destination vector of the route will be set for the destination VCPU.

The intention is for the userspace IOAPICs to use the reservable MSI
routes to inject interrupts into the guest.

This is a slight abuse of the notion of an MSI Route, given that MSIs
classically bypass the IOAPIC. It might be worthwhile to add an
additional route type to improve clarity.

Compile tested for Intel x86.

Signed-off-by: Steve Rutherford <srutherford@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[RHEL: out of order backport of IOAPIC scan fixes and disabled_lapic_foudn
 (from the vector hashing patches)]
(cherry picked from commit b053b2aef25d00773fa6762dcd4b7f5c9c42d171)

Signed-off-by: Rafael Aquini <aquini@redhat.com>