[virt] kvm: x86: request interrupt window when IRQ chip is split
Message-id: <1467818058-27383-67-git-send-email-pbonzini@redhat.com>
Patchwork-id: 154755
O-Subject: [RHEL7.3 PATCH v3 66/77] KVM: x86: request interrupt window when IRQ chip is split
Bugzilla: 1347370
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Matt Gingell <gingell@google.com>

Before this patch, we incorrectly enter the guest without requesting an
interrupt window if the IRQ chip is split between user space and the
kernel.

Because lapic_in_kernel no longer implies the PIC is in the kernel, this
patch tests pic_in_kernel to determining whether an interrupt window
should be requested when entering the guest.

If the APIC is in the kernel and we request an interrupt window the
guest will return immediately. If the APIC is masked the guest will not
not make forward progress and unmask it, leading to a loop when KVM
reenters and requests again. This patch adds a check to ensure the APIC
is ready to accept an interrupt before requesting a window.

Reviewed-by: Steve Rutherford <srutherford@google.com>
Signed-off-by: Matt Gingell <gingell@google.com>
[Use the other newly introduced functions. - Paolo]
Fixes: 1c1a9ce973a7863dd46767226bce2a5f12d48bc6
Cc: stable@vger.kernel.org
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

(cherry picked from commit 62a193edaf90df38356e292f47a17f28e0cee3f1)
Signed-off-by: Rafael Aquini <aquini@redhat.com>