[fs] gfs2: Check rs_free with rd_rsspin protection
Message-id: <298288819.7051743.1467820959552.JavaMail.zimbra@redhat.com>
Patchwork-id: 154768
O-Subject: [RHEL7.3 fs PATCH] GFS2: Check rs_free with rd_rsspin protection
Bugzilla: 1349596
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bug 1349596 - GFS2: kernel BUG at fs/gfs2/rgrp.c:649!

The patch was posted upstream, but has NOT yet been accepted nor ACKed.
I'm posting it anyway due to the deadline.

It was built into a RHEL7.3 kernel here:

https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11308302

It was tested by Barry using specsfs, which was consistently
failing before.

Patch description:

For the last process to close a file opened for write, function
gfs2_rsqa_delete was deleting the file's inode's block reservation
out of the rgrp reservations tree. Then it was checking to make sure
rs_free was 0, but it was performing the check outside the protection
of rd_rsspin spin_lock. The rd_rsspin spin_lock protection is needed
to prevent a race between the process freeing the reservation and
another who is allocating a new set of blocks inside the same rgrp
for the same inode, thus changing its value.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1349596
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>