[fs] gfs2: writeout truncated pages
Message-id: <1467840764-969-3-git-send-email-bmarzins@redhat.com>
Patchwork-id: 154841
O-Subject: [RHEL7.3 BZ1221210 PATCH 2/2] gfs2: writeout truncated pages
Bugzilla: 1221210
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1221210

When gfs2 attempts to write a page to a file that is being truncated,
and notices that the page is completely outside of the file size, it
tries to invalidate it.  However, this may require a transaction for
journaled data files to revoke any buffers from the page on the active
items list. Unfortunately, this can happen inside a log flush, where a
transaction cannot be started. Also, gfs2 may need to be able to remove
the buffer from the ail1 list before it can finish the log flush.

To deal with this, when writing a page of a file with data journalling
enabled gfs2 now skips the check to see if the write is outside the file
size, and simply writes it anyway. This situation can only occur when
the truncate code still has the file locked exclusively, and hasn't
marked this block as free in the metadata (which happens later in
truc_dealloc).  After gfs2 writes this page out, the truncation code
will shortly invalidate it and write out any revokes if necessary.

To do this, gfs2 now implements its own version of block_write_full_page
without the check, and calls the newly exported __block_write_full_page.
It also no longer calls gfs2_writepage_common. Instead the necessary
checks from there have been pulled into gfs2_jdata_writepage

This patch is already in the for-next branch of the linux-gfs2 upstream
git repository:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/?h=for-next&id=fd4c5748b8d3f7420e8932ed0bde3d53cc8acc9d

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>