[kernel] introduce tasklist_read_lock() and tasklist_write_lock_irq()
Message-id: <20160708170103.GA26970@redhat.com>
Patchwork-id: 154991
O-Subject: [PATCH RHEL7.3 BZ 1243748 1/3] introduce tasklist_read_lock() and tasklist_write_lock_irq()
Bugzilla: 1243748
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

RHBZ: 1243748

RHEL-only. Comes as a separate patch to simplify the review.

tasklist_lock is rwlock_t and thus it is unfair, the new reader blocks the
pending writers. Unfortunately it is very easy to trigger read_lock() from
user-space and make the writers starve. For example I was able to lock-up
the (idle!) 8 CPU machine just doing

	$ perl -e 'fork for 1..4; kill SIGWINCH, -1 while 1'

and
	$ perl -le 'fork ? print wait : exit while 1'

on another console. These kill(SIGWINCH) from 16 processes in parallel block
the writes completely, so that fork/exit/etc can't succeed.

This was fixed upstream (although I believe only partially fixed), rwlock_t
was turned into the fair qrwlock. But I do not think we can do this in rhel7
at least until we have other reasons to backport the queued spin/rw-locks.

And as for tasklist_lock in particular I have other concerns which make me
worry. tasklist_lock won't be recursive any longer, the nested readers can
deadlock (unless the 2nd reader comes from irq) and lockdep can not detect
the problem. This also affects the lock ordering, for example tasklist_lock
no longer  can nest inside task_lock() despite the fact this is documented
in sched.h.

So the best solution I can suggest for now is this ugly hack which turns
tasklist_lock into the write-biased lock. Ugly but simple, we will probably
need to backport this fix to rhel7.2 as was already requested by Fujitsu.

Performance-wise this change is obviously unfortunate. _write() is probably
fine, if write_trylock() fails atomic_inc() should not hurt because we are
going to spin in write_lock() which writes to the sane cache-line anyway.
But tasklist_read_lock() does the unconditional LOAD before read_lock() and
this doesn't look nice. I _hope_ this is fine, we have a lot more performance
problems with tasklisk_lock which does not scale at all. Plus the only user
of tasklist_read_lock() which we do care about (performance-wise) is do_wait(),
and it is heavy enough either way.

However I do not really know how to verify/prove that this won't cause any
kind of performance regression. I played with "make -j bzImage" and did not
notice any meaningful difference, but that is all.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>