[fs] cifs: Check for existing directory when opening file with O_CREAT
Message-id: <1468408551-18122-1-git-send-email-sprabhu@redhat.com>
Patchwork-id: 155107
O-Subject: [RHEL7 PATCH fs] [bz1346118]cifs: Check for existing directory when opening file with O_CREAT
Bugzilla: 1346118
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

bz: 1346118

This has been tested successfully against the reproducer from the bz.

The bug allows a non privileged user to crash a node which has mounted a
SMB1 share by opening a directory whoose attributes haven't been cached
with the O_CREAT flag. When attempting to open a directory with the
O_CREAT flag, we now return -EISDIR in the same manner as local
filesystems.

The patch has already been sent upstream. This is currently in Steve
French's git tree.
https://git.samba.org/?p=sfrench/cifs-2.6.git;a=commit;h=8d9535b6efd86e6c07da59f97e68f44efb7fe080

Original summary:

When opening a file with O_CREAT flag, check to see if the file opened
is an existing directory.

This prevents the directory from being opened which subsequently causes
a crash when the close function for directories cifs_closedir() is called
which frees up the file->private_data memory while the file is still
listed on the open file list for the tcon.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Reported-by: Xiaoli Feng <xifeng@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>