[fs] ext4: add remap_file_pages support for dax mounts
Message-id: <20160715184552.13004-3-jmoyer@redhat.com>
Patchwork-id: 155491
O-Subject: [RHEL7 PATCH 2/3] ext4: add remap_file_pages support for dax mounts
Bugzilla: 1348428
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

BZ: 1348428

The original patch adding DAX functionality to ext4 filled in a
remap_pages function pointer:
  https://lkml.org/lkml/2014/2/25/487

However, since remap_pages was removed by the time the patch was
merged, that bit got dropped.  Our backport was based on the patch
that got merged, so we missed this.  As a result, remap_file_pages
fails on ext4 file systems mounted with "-o dax".  Fixing this in
our kernel is as easy as filling that function pointer back in.

This patch is RHEL-only, obviously.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>