[fs] cifs: remove any preceding delimiter from prefix_path
Message-id: <1468856702-2388-1-git-send-email-sprabhu@redhat.com>
Patchwork-id: 155541
O-Subject: [RHEL7 PATCH fs 1/2] [bz1252721]cifs: remove any preceding delimiter from prefix_path
Bugzilla: 1252721
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Bz: 1252721
Upstream commit: 11e31647c9263185cfd990db656dbb7f06697faf
Brew build:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11363989

Original summary
--
We currently do not check if any delimiter exists before the prefix
path in cifs_compose_mount_options(). Consequently when building the
devname using cifs_build_devname() we can end up with multiple
delimiters separating the UNC and the prefix path.

An issue was reported by the customer mounting a folder within a DFS
share from a Netapp server which uses McAfee antivirus. We have
narrowed down the cause to the use of double backslashes in the file
name used to open the file. This was determined to be caused because of
additional delimiters as a result of the bug.

In addition to changes in cifs_build_devname(), we also fix
cifs_parse_devname() to ignore any preceding delimiter for the prefix
path.

The problem was originally reported on RHEL 6 in RHEL bz 1252721. This
is the upstream version of the fix. The fix was confirmed by looking at
the packet capture of a DFS mount.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 11e31647c9263185cfd990db656dbb7f06697faf)
--

Tested successfully against the reproducer in the bz.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>