[infiniband] ib/core: Make device counter infrastructure dynamic
Message-id: <1468984921-30188-2-git-send-email-ddutile@redhat.com>
Patchwork-id: 155713
O-Subject: [PATCH RHEL-7.3 1/9 v2] IB/core: Make device counter infrastructure dynamic
Bugzilla: 1258655
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Christoph Lameter <cl@linux.com>

In practice, each RDMA device has a unique set of counters that the
hardware implements.  Having a central set of counters that they must
all adhere to is limiting and causes many useful counters to not be
available.

Therefore we create a dynamic counter registration infrastructure.

The driver must implement a stats structure allocation routine, in
which the driver must place the directory name it wants, a list of
names for all of the counters, an array of u64 counters themselves,
plus a few generic configuration options.

We then implement a core routine to create a sysfs file for each
of the named stats elements, and a core routine to retrieve the
stats when any of the sysfs attribute files are read.

To avoid excessive beating on the stats generation routine in the
drivers, the core code also caches the stats for a short period of
time so that someone attempting to read all of the stats in a
given device's directory will not result in a stats generation
call per file read.

Future work will attempt to standardize just the shared stats
elements, and possibly add a method to get the stats via netlink
in addition to sysfs.

Signed-off-by: Christoph Lameter <cl@linux.com>
Signed-off-by: Mark Bloch <markb@mellanox.com>
Reviewed-by: Steve Wise <swise@opengridcomputing.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
[ Add caching, make structure names more informative, add i40iw support,
  other significant rewrites from the original patch ]
(cherry picked from commit b40f4757daa1b28e586fddad76638c98e2edfc34)

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/infiniband/hw/i40iw/i40iw_verbs.c
Not in RHEL-7.3 yet, so not this hunk of upstream commit dropped.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1258655
Brew build: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=11378457

Testing:
Customer making request for feature tested -471-based kernel with patch.

Signed-off-by: Donald Dutile <ddutile@redhat.com>