[target] iscsi-target: Rework the configfs of cxgbit
Message-id: <1469051220-25779-32-git-send-email-svemuri@redhat.com>
Patchwork-id: 155757
O-Subject: [RHEL7.3 BZ1211351 PATCH 31/31] iscsi-target : Rework the configfs of cxgbit
Bugzilla: 1211351
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Andy Grover <agrover@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla - https://bugzilla.redhat.com/show_bug.cgi?id=1211351

This is a RHEL only patch.

There was a major change in upstream lio-target code because of commit,
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=2eafd72939fda6118e27d3ee859684987f43921b.

The cxgbit changes sit on top of this change. Since the commit mentioned above is not there in RHEL,
we have changed cxgbit offload registration to follow the old style.

This patch fixes the issues related to Chelsio's LIO iscsi-target mentioned in BZ 1354412.

Signed-off-by: Rafael Aquini <aquini@redhat.com>