[fs] xfs: fix duplicate buffer flag bits
Message-id: <1469110743-62086-2-git-send-email-bfoster@redhat.com>
Patchwork-id: 155807
O-Subject: [PATCH RHEL7 1/3] xfs: fix duplicate buffer flag bits
Bugzilla: 1347744 1358817
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

XBF_WRITE_FAIL and _XBF_LRU_DISPOSE flags are both defined to bit 24 of
the buffer flags field. This is the result of a backport error in rhel7
commit 9ef79444 ("xfs: abort metadata writeback on permanent errors").

This occurred because _XBF_LRU_DISPOSE was originally defined to use bit
24 and backported as such. It was then removed upstream by the time
XBF_WRITE_FAIL was introduced, which reused bit 24. _XBF_LRU_DISPOSE
remained in rhel7, however, as the upstream removal was part of a larger
lru rework that was not backported. Thus, XBF_WRITE_FAIL was mistakenly
backported to reuse the same bit in rhel7.

Fix up _XBF_WRITE_FAIL to use a unique buffer flags bit. This is a
RHEL-specific problem and thus a RHEL-specific patch.

Signed-off-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>