[crypto] x86/sha512_ssse3 - fixup for asm function prototype change
Message-id: <20160724002902.GA30430@gondor.apana.org.au>
Patchwork-id: 155886
O-Subject: [RHEL7 PATCH] crypto: x86/sha512_ssse3 - fixup for asm function prototype change
Bugzilla: 1267049
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHEL7 Bugzilla #1267049

Upstream commit: 00425bb181c204c8f250fec122e2817a930e0286

Patch e68410ebf626 ("crypto: x86/sha512_ssse3 - move SHA-384/512
SSSE3 implementation to base layer") changed the prototypes of the
core asm SHA-512 implementations so that they are compatible with
the prototype used by the base layer.

However, in one instance, the register that was used for passing the
input buffer was reused as a scratch register later on in the code,
and since the input buffer param changed places with the digest param
-which needs to be written back before the function returns- this
resulted in the scratch register to be dereferenced in a memory write
operation, causing a GPF.

Fix this by changing the scratch register to use the same register as
the input buffer param again.

Fixes: e68410ebf626 ("crypto: x86/sha512_ssse3 - move SHA-384/512 SSSE3 implementation to base layer")
Reported-By: Bobby Powers <bobbypowers@gmail.com>
Tested-By: Bobby Powers <bobbypowers@gmail.com>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Rafael Aquini <aquini@redhat.com>