[fs] gfs2: Add meta readahead field in directory entries
Message-id: <1469480565-20272-2-git-send-email-agruenba@redhat.com>
Patchwork-id: 155947
O-Subject: [RHEL7.4 1/3] [fs] GFS2: Add meta readahead field in directory entries
Bugzilla: 1256539
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

Bugzilla: 1256539
Upstream status: This is a backport of upstream's commit 44aaada9d1.

The intent of this new field in the directory entry is to
allow a subsequent lookup to know how many blocks, which
are contiguous with the inode, contain metadata which relates
to the inode. This will then allow the issuing of a single
read to read these blocks, rather than reading the inode
first, and then issuing a second read for the metadata.

This only works under some fairly strict conditions, since
we do not have back pointers from inodes to directory entries
we must ensure that the blocks referenced in this way will
always belong to the inode.

This rules out being able to use this system for indirect
blocks, as these can change as a result of truncate/rewrite.

So the idea here is to restrict this to xattr blocks only
for the time being. For most inodes, that means only a
single block. Also, when using ACLs and/or SELinux or
other LSMs, these will be added at inode creation time
so that they will be contiguous with the inode on disk and
also will almost always be needed when we read the inode in
for permissions checks.

Once an xattr block for an inode is allocated, it will never
change until the inode is deallocated.

This patch adds the new field, a further patch will add the
readahead in due course.

(Note: the actual declaration of de_rahead in struct gfs2_dirent
was already backported in commit 3169fd72dc.)

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>