[fs] gfs2: Extended attribute readahead
Message-id: <1469480565-20272-3-git-send-email-agruenba@redhat.com>
Patchwork-id: 155948
O-Subject: [RHEL7.4 2/3] [fs] gfs2: Extended attribute readahead
Bugzilla: 1256539
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Bugzilla: 1256539
Upstream Status: This is a backport of upstream's commit c8d5770384.

When gfs2 allocates an inode and its extended attribute block next to
each other at inode create time, the inode's directory entry indicates
that in de_rahead.  In that case, we can readahead the extended
attribute block when we read in the inode.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>