[security] keys: Don't leak a key reference if request_key() tries to use a revoked keyring
Message-id: <146954584277.28870.14496952638277135109.stgit@warthog.procyon.org.uk>
Patchwork-id: 155980
O-Subject: [RHEL7 PATCH] Don't leak a key reference if request_key() tries to use a revoked keyring
Bugzilla: 1282584
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: David Jeffery <djeffery@redhat.com>

If a request_key() call to allocate and fill out a key attempts to insert the
key structure into a revoked keyring, the key will leak, using memory and part
of the user's key quota until the system reboots. This is from a failure of
construct_alloc_key() to decrement the key's reference count after the attempt
to insert into the requested keyring is rejected.

key_put() needs to be called in the link_prealloc_failed callpath to ensure
the unused key is released.

Signed-off-by: David Jeffery <djeffery@redhat.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: James Morris <james.l.morris@oracle.com>
Original-commit: d0709f1e66e8066c4ac6a54620ec116aa41937c0
RH-Bugzilla: 1282584
Signed-off-by: Rafael Aquini <aquini@redhat.com>